% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{catalogue.compile}
\alias{catalogue.compile}
\title{TBD - temporarily export-ed}
\usage{
catalogue.compile(fishery_ranks_data, catalogue_data)
}
\arguments{
\item{fishery_ranks_data}{TBD}

\item{catalogue_data}{TBD}
}
\value{
TBD
}
\description{
TBD - temporarily export-ed
}
