% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_AREAS}
\alias{REF_AREAS}
\title{The reference fishing areas}
\format{
\describe{
  \item{CODE}{The fishing zone code}
  \item{QUADRANT_CODE}{The fishing area quadrant code}
  \item{GEO_AREA_CODE}{The fishing area geographical area code}
  \item{NAME_EN}{The fishing area English name}
  \item{DEPRECATED}{If the fishing area is deprecated and kept for historical reasons only}
}
}
\usage{
REF_AREAS
}
\description{
The reference fishing areas
}
\keyword{datasets}
