% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_CATCH_TYPES}
\alias{REF_CATCH_TYPES}
\title{The reference catch types}
\format{
\describe{
  \item{ID}{The catch type ID}
  \item{CODE}{The catch type code}
  \item{NAME_EN}{The catch type English name}
  \item{NAME_ES}{The catch type Spanish name}
  \item{NAME_FR}{The catch type French name}
  \item{CATCH_TYPE_GROUP}{The catch type group}
}
}
\usage{
REF_CATCH_TYPES
}
\description{
The reference catch types
}
\keyword{datasets}
