% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_DATA_SOURCE_CONTENTS}
\alias{REF_DATA_SOURCE_CONTENTS}
\title{The data source contents}
\format{
\describe{
  \item{CODE}{The data source content code}
  \item{NAME_EN}{The data source content English name}
  \item{NAME_ES}{The data source content Spanish name}
  \item{NAME_FR}{The data source content French name}
}
}
\usage{
REF_DATA_SOURCE_CONTENTS
}
\description{
The data source contents
}
\keyword{datasets}
