% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_DATA_SOURCES}
\alias{REF_DATA_SOURCES}
\title{The data sources}
\format{
\describe{
  \item{CODE}{The data source code}
  \item{NAME_EN}{The data source English name}
}
}
\usage{
REF_DATA_SOURCES
}
\description{
The data sources
}
\keyword{datasets}
