% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_EFFORT_TYPES}
\alias{REF_EFFORT_TYPES}
\title{The reference effort types}
\format{
\describe{
  \item{CODE}{The effort type code}
  \item{NAME_EN}{The effort type English name}
  \item{GEAR_DEPENDENT}{If the effort type is associated to a specific gear}
}
}
\usage{
REF_EFFORT_TYPES
}
\description{
The reference effort types
}
\keyword{datasets}
