% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_FISHING_ZONES}
\alias{REF_FISHING_ZONES}
\title{The reference fishing zones}
\format{
\describe{
  \item{CODE}{The fishing zone code}
  \item{NAME_EN}{The fishing zone English name}
}
}
\usage{
REF_FISHING_ZONES
}
\description{
The reference fishing zones
}
\keyword{datasets}
