% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_FLAGS}
\alias{REF_FLAGS}
\title{The reference flags}
\format{
\describe{
  \item{CODE}{The flag code}
  \item{NAME_EN}{The flag English name}
}
}
\usage{
REF_FLAGS
}
\description{
The reference flags
}
\keyword{datasets}
