% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_FLEETS}
\alias{REF_FLEETS}
\title{The reference fleets}
\format{
\describe{
  \item{CODE}{The fleet code}
  \item{FLAG_CODE}{The flag code for the fleet}
  \item{VESSEL_FLAG_CODE}{The vessel flag code for the fleet}
  \item{NAME_EN}{The fleet English name}
  \item{DEPRECATED}{If the fleet is deprecated and kept for historical reasons only}
}
}
\usage{
REF_FLEETS
}
\description{
The reference fleets
}
\keyword{datasets}
