% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_FREQUENCY_TYPES}
\alias{REF_FREQUENCY_TYPES}
\title{The frequency types}
\format{
\describe{
  \item{CODE}{The frequency type code}
  \item{NAME_EN}{The frequency type English name}
  \item{FREQUENCY_TYPE_GROUP_CODE}{The frequency type group code}
}
}
\usage{
REF_FREQUENCY_TYPES
}
\description{
The frequency types
}
\keyword{datasets}
