% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_GEAR_GROUPS}
\alias{REF_GEAR_GROUPS}
\title{The reference gear groups}
\format{
\describe{
  \item{CODE}{The gear group code}
  \item{NAME_EN}{The gear group English name}
}
}
\usage{
REF_GEAR_GROUPS
}
\description{
The reference gear groups
}
\keyword{datasets}
