% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_GEARS}
\alias{REF_GEARS}
\title{The reference gears}
\format{
\describe{
  \item{CODE}{The gear code}
  \item{GEAR_GROUP_CODE}{The gear group code}
  \item{NAME_EN}{The gear English name}
  \item{DISCARDS}{If the specific gear is used to report discards - DEPRECATED}
}
}
\usage{
REF_GEARS
}
\description{
The reference gears
}
\keyword{datasets}
