% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_PARTIES}
\alias{REF_PARTIES}
\title{The reference parties (CPCs)}
\format{
\describe{
  \item{CODE}{The party code}
  \item{STATUS_TYPE_CODE}{The party status code}
  \item{ENTITY_TYPE_CODE}{The party entity code}
  \item{NAME_EN}{The party English name}
  \item{ACCESSION_DATE}{The party accession date}
}
}
\usage{
REF_PARTIES
}
\description{
The reference parties (CPCs)
}
\keyword{datasets}
