% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_PARTY_STATUS}
\alias{REF_PARTY_STATUS}
\title{The reference party status}
\format{
\describe{
  \item{CODE}{The party status code}
  \item{NAME_EN}{The party status English name}
  \item{DESCRIPTION_EN}{The party status English description}
}
}
\usage{
REF_PARTY_STATUS
}
\description{
The reference party status
}
\keyword{datasets}
