% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_PRODUCT_TYPES}
\alias{REF_PRODUCT_TYPES}
\title{The product types}
\format{
\describe{
  \item{CODE}{The product type code}
  \item{NAME_EN}{The product type English name}
  \item{IS_WEIGHT}{If the product type refers to a weight (1) or to a number (0)}
}
}
\usage{
REF_PRODUCT_TYPES
}
\description{
The product types
}
\keyword{datasets}
