% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_QUALITY_LEVELS}
\alias{REF_QUALITY_LEVELS}
\title{The reference data quality levels}
\format{
\describe{
  \item{CODE}{The quality level code}
  \item{QUALITY_GROUP_CODE}{The quality level group code}
  \item{NAME_EN}{The quality level English name}
  \item{DESCRIPTION_EN}{The quality level English description}
}
}
\usage{
REF_QUALITY_LEVELS
}
\description{
The reference data quality levels
}
\keyword{datasets}
