% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_SAMPLING_AREAS}
\alias{REF_SAMPLING_AREAS}
\title{The reference sampling areas}
\format{
\describe{
  \item{CODE}{The sampling area code}
  \item{NAME_EN}{The sampling area English name}
}
}
\usage{
REF_SAMPLING_AREAS
}
\description{
The reference sampling areas
}
\keyword{datasets}
