% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_SAMPLING_LOCATIONS}
\alias{REF_SAMPLING_LOCATIONS}
\title{The sampling locations}
\format{
\describe{
  \item{CODE}{The sampling location code}
  \item{NAME_EN}{The sampling location English name}
}
}
\usage{
REF_SAMPLING_LOCATIONS
}
\description{
The sampling locations
}
\keyword{datasets}
