% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_SAMPLING_UNITS}
\alias{REF_SAMPLING_UNITS}
\title{The sampling units}
\format{
\describe{
  \item{CODE}{The sampling unit code}
  \item{NAME_EN}{The sampling unit English name}
}
}
\usage{
REF_SAMPLING_UNITS
}
\description{
The sampling units
}
\keyword{datasets}
