% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_SIZE_CLASS_LIMITS}
\alias{REF_SIZE_CLASS_LIMITS}
\title{The reference size class limits}
\format{
\describe{
  \item{CODE}{The size class limit code}
  \item{NAME_EN}{The size class limit English name}
}
}
\usage{
REF_SIZE_CLASS_LIMITS
}
\description{
The reference size class limits
}
\keyword{datasets}
