% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_SPECIES}
\alias{REF_SPECIES}
\title{The reference species}
\format{
\describe{
  \item{ID}{The species ID (ICCAT)}
  \item{CODE}{The species code (FAO 3-ALPHA)}
  \item{SCIENTIFIC_NAME}{The species scientific name}
  \item{NAME_EN}{The species English name}
  \item{NAME_ES}{The species Spanish name}
  \item{NAME_FR}{The species French name}
  \item{SPECIES_GROUP}{The species group}
  \item{SPECIES_GROUP_ICCAT}{The ICCAT species group}
}
}
\usage{
REF_SPECIES
}
\description{
The reference species
}
\keyword{datasets}
