% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_SPECIES_GEAR_GROUPS}
\alias{REF_SPECIES_GEAR_GROUPS}
\title{The reference gear groups by species}
\format{
\describe{
  \item{SPECIES_CODE}{The species code}
  \item{GEAR_GROUP_CODE}{The gear group code}
  \item{SPECIES_GEAR_GROUP}{The species gear group}
  \item{SPECIES_GEAR_GROUP_ORDER}{The species gear group order}
}
}
\usage{
REF_SPECIES_GEAR_GROUPS
}
\description{
The reference gear groups by species
}
\keyword{datasets}
