% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_SPECIES_META}
\alias{REF_SPECIES_META}
\title{The reference species metadata}
\format{
\describe{
  \item{CODE}{The species code}
  \item{NAME_EN}{The species English name}
  \item{NAME_ES}{The species Spanish name}
  \item{NAME_FR}{The species French name}
  \item{STOCK_BOUNDARY_ID}{The species stock boundary ID}
  \item{SPECIES_GROUP_CODE}{The species group code}
}
}
\usage{
REF_SPECIES_META
}
\description{
The reference species metadata
}
\keyword{datasets}
