% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_SQUARE_TYPES}
\alias{REF_SQUARE_TYPES}
\title{The reference square types}
\format{
\describe{
  \item{CODE}{The square type code}
  \item{NAME_EN}{The square type English name}
  \item{DEPRECATED}{If the square type is deprecated and kept for historical reasons only}
}
}
\usage{
REF_SQUARE_TYPES
}
\description{
The reference square types
}
\keyword{datasets}
