% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_STOCK_AREAS}
\alias{REF_STOCK_AREAS}
\title{The reference stock areas}
\format{
\describe{
  \item{CODE}{The stock area code}
  \item{NAME_EN}{The stock area English name}
}
}
\usage{
REF_STOCK_AREAS
}
\description{
The reference stock areas
}
\keyword{datasets}
