% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_STOCKS}
\alias{REF_STOCKS}
\title{The reference stock metadata}
\format{
\describe{
  \item{SPECIES_CODE}{The species code}
  \item{SPECIES_GROUP}{The species group}
  \item{STOCK_CODE}{The stock code}
  \item{STATISTICAL_AREA_CODE}{The statistical area code}
  \item{SAMPLING_AREA_CODE}{The sampling area code}
}
}
\usage{
REF_STOCKS
}
\description{
The reference stock metadata
}
\keyword{datasets}
