% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_TIME_PERIOD_GROUPS}
\alias{REF_TIME_PERIOD_GROUPS}
\title{The reference time period groups}
\format{
\describe{
  \item{CODE}{The time period group code}
  \item{NAME_EN}{The time period group English name}
}
}
\usage{
REF_TIME_PERIOD_GROUPS
}
\description{
The reference time period groups
}
\keyword{datasets}
