% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REF_TIME_PERIODS}
\alias{REF_TIME_PERIODS}
\title{The reference time periods}
\format{
\describe{
  \item{CODE}{The time period code}
  \item{TIME_PERIOD_GROUP_CODE}{The time period group code}
  \item{NAME_EN}{The time period English name}
}
}
\usage{
REF_TIME_PERIODS
}
\description{
The reference time periods
}
\keyword{datasets}
