% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc.R
\name{t1nc.summarise}
\alias{t1nc.summarise}
\title{Produces a summary (wide) table of nominal catches grouped by year and stratified by
several criteria, such as species, stock, gear, catch type, with the possible inclusion
of fishery ranks for each stratum}
\usage{
t1nc.summarise(
  t1nc_data,
  year_min = NA,
  year_max = NA,
  by_species = TRUE,
  by_stock = TRUE,
  by_gear = TRUE,
  by_catch_type = TRUE,
  rank = FALSE
)
}
\arguments{
\item{t1nc_data}{the original T1NC data, in long form, and with the same structure as the data that can be retrieved using the \code{\link{iccat.dev.data::t1nc}} function of the iccat.dev.data library}

\item{year_min}{the starting year to be included in the output}

\item{year_max}{the ending year to be included in the output}

\item{by_species}{whether or not the output stratification shall include species codes}

\item{by_stock}{whether or not the output stratification shall include stock codes}

\item{by_gear}{whether or not the output stratification shall include gear group codes}

\item{by_catch_type}{whether or not the output stratification shall include catch type codes}

\item{rank}{whether or not to produce the average catch ratio (absolute / cumulative) for each strata}
}
\value{
a wide table of nominal catches grouped by year and stratified according to the chosen criteria
}
\description{
Produces a summary (wide) table of nominal catches grouped by year and stratified by
several criteria, such as species, stock, gear, catch type, with the possible inclusion
of fishery ranks for each stratum
}
