% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{catalogue.compile}
\alias{catalogue.compile}
\title{TBD - temporarily export-ed}
\usage{
catalogue.compile(
  fishery_ranks_data,
  catalogue_data,
  year_from = NA,
  year_to = NA,
  pretty_print_catches = TRUE,
  catch_round_digits = 0,
  perc_round_digits = 2
)
}
\arguments{
\item{fishery_ranks_data}{TBD}

\item{catalogue_data}{TBD}

\item{year_from}{TBD}

\item{year_to}{TBD}

\item{pretty_print_catches}{TBD}

\item{catch_round_digits}{TBD}

\item{perc_round_digits}{TBD}
}
\value{
TBD
}
\description{
TBD - temporarily export-ed
}
