% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc.R
\name{t1nc.summarise}
\alias{t1nc.summarise}
\title{TBD}
\usage{
t1nc.summarise(t1nc_data, by_species = TRUE, by_stock = TRUE, by_gear = TRUE)
}
\arguments{
\item{t1nc_data}{TBD}

\item{by_species}{TBD}

\item{by_stock}{TBD}

\item{by_gear}{TBD}
}
\value{
TBD
}
\description{
TBD
}
