% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1nc.R
\name{t1nc.summarise}
\alias{t1nc.summarise}
\title{TBD}
\usage{
t1nc.summarise(
  t1nc_data,
  year_min = NA,
  year_max = NA,
  by_species = TRUE,
  by_stock = TRUE,
  by_gear = TRUE,
  by_catch_type = TRUE,
  rank = FALSE
)
}
\arguments{
\item{t1nc_data}{TBD}

\item{by_species}{TBD}

\item{by_stock}{TBD}

\item{by_gear}{TBD}

\item{by_catch_type}{TBD}

\item{rank}{TBD}
}
\value{
TBD
}
\description{
TBD
}
